services/postgres/requirements.txt
appdirs==1.4.3
astroid==2.4.0
attrs==19.3.0
black==19.10b0
click==7.1.2
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pathspec==0.8.0
psycopg2-binary==2.8.5
pylint==2.5.0
regex==2020.4.4
sentry-sdk==0.14.4
six==1.14.0
toml==0.10.0
typed-ast==1.4.1
wrapt==1.12.1
